Fix the def integration for ECL.
This problem is similar to issues related to in-package
handling wrt hu.dwim.common. Namely, unterning the symbol
during compilation causes the compiler to externalize
unexport's argument as #:-body-, thus causing it to break
at load time. This was compounded by a bug in ECL's error
checking that let unexport succeed, silently corrupting
internal state, and causing a crash later on.